+++ src/prelude.rs
@@ -1,0 +1,6 @@
+// TODO: remove
+
+//pub use crate::BitSet;
+pub use crate::BitSetInterface;
+pub use crate::iter::IndexIterator;    
+pub use crate::iter::BlockIterator;