+++ src/test.rs
@@ -105,6 +105,7 @@ fn fuzzy_test(){
         const INDEX_MUL: usize = 10;
     }
     }
+    const MAX_CURSOR_READ_SESSION: usize = MAX_SIZE;
 
     let mut rng = rand::thread_rng();
     for _ in 0..REPEATS{
@@ -147,7 +148,7 @@ fn fuzzy_test(){
                 assert!(hi_set.contains(index));
             }
             
-            // block traverse
+            // block traverse contains
             hi_set.block_iter().traverse(|block|{ 
                 block.traverse(|index|{
                     assert!(hash_set.contains(&index));
@@ -155,7 +156,7 @@ fn fuzzy_test(){
                 })
             });
 
-            // index traverse
+            // index traverse contains
             hi_set.iter().for_each(|index|{ 
                 assert!(hash_set.contains(&index));
             });            
@@ -171,6 +172,59 @@ fn fuzzy_test(){
             {
                 let other: HiSparseBitset = hi_set.iter().collect();
                 assert!(hi_set == other);
+            }
+
+            let mut hash_set_vec: Vec<usize> = hash_set.iter().copied().collect();
+            hash_set_vec.sort();
+            
+            // block traverse cursor sessions
+            {
+                let mut cursor = BlockCursor::start();
+                let mut traversed = Vec::new();
+
+                loop{
+                    let mut session_counter = rng.gen_range(0..MAX_CURSOR_READ_SESSION) as isize;                
+                    let ctrl = hi_set.block_iter().move_to(cursor).traverse(|block|{
+                        if session_counter <= 0{
+                            cursor = (&block).into();
+                            return ControlFlow::Break(());
+                        }
+                        session_counter -= block.len() as isize;
+
+                        traversed.extend(block);
+                        ControlFlow::Continue(())
+                    });
+                    if ctrl.is_continue(){
+                        break;
+                    }
+                }
+
+                assert_equal(traversed, hash_set_vec.iter().copied());
+            }
+            
+            // index traverse cursor sessions
+            {
+                let mut cursor = IndexCursor::start();
+                let mut traversed = Vec::new();
+
+                loop{
+                    let mut session_counter = rng.gen_range(0..MAX_CURSOR_READ_SESSION);
+                    let ctrl = hi_set.iter().move_to(cursor).traverse(|index|{
+                        if session_counter == 0{
+                            cursor = index.into();
+                            return ControlFlow::Break(());
+                        }
+                        session_counter -= 1;
+
+                        traversed.push(index);
+                        ControlFlow::Continue(())
+                    });
+                    if ctrl.is_continue(){
+                        break;
+                    }
+                }
+
+                assert_equal(traversed, hash_set_vec.iter().copied());
             }
         }
     }
@@ -310,15 +364,17 @@ where
                 let mut blocks_to_consume = rng.gen_range(0..MAX_RESUMED_INTERSECTION_BLOCKS_CONSUME);
 
                 // through traverse
+                let mut traversed_cursor = BlockCursor::end();
                 let mut traversed_blocks = Vec::new();
                 {
                     let mut blocks_to_consume = blocks_to_consume;
                     intersection.clone().traverse(|block|{
                         if blocks_to_consume == 0{
+                            traversed_cursor = BlockCursor::from(&block);
                             return ControlFlow::Break(());
                         }
                         blocks_to_consume -= 1;
-
+                                                 
                         traversed_blocks.push(block);
                         ControlFlow::Continue(())
                     });
@@ -353,6 +409,10 @@ where
                 assert_equal(traversed_blocks, iterated_blocks);
 
                 block_cursor = intersection.cursor();
+                assert_eq!(
+                    intersection.clone().move_to(block_cursor).next(),
+                    intersection.clone().move_to(traversed_cursor).next()
+                );
             }
 
             // suspend/resume indices
@@ -365,11 +425,13 @@ where
                 let indices_to_consume = rng.gen_range(0..MAX_RESUMED_INTERSECTION_INDICES_CONSUME);
 
                 // through traverse
+                let mut traversed_cursor = IndexCursor::end();
                 let mut traversed_indices = Vec::new();
                 {
                     let mut indices_to_consume = indices_to_consume;
                     intersection.clone().traverse(|i|{
                         if indices_to_consume == 0{
+                            traversed_cursor = i.into();
                             return ControlFlow::Break(());
                         }
                         indices_to_consume -= 1;
@@ -379,6 +441,7 @@ where
                     });
                 }
 
+                // iteration
                 let mut iterated_indices = Vec::new();
                 for index in intersection.by_ref().take(indices_to_consume){
                     assert!(hashsets_intersection.contains(&index));
@@ -389,8 +452,12 @@ where
                     iterated_indices.push(index);
                 }
 
-                assert_equal(traversed_indices, iterated_indices);
                 index_cursor = intersection.cursor();
+                assert_equal(traversed_indices, iterated_indices);
+                assert_eq!(
+                    intersection.clone().move_to(index_cursor).next(),
+                    intersection.clone().move_to(traversed_cursor).next()
+                );
             }
 
             // reduce test
@@ -882,7 +949,6 @@ fn block_cursor_test_empty2(){
     assert!(iter.next().is_none());
 }
 
-
 #[test]
 fn index_cursor_test(){
     type HiSparseBitset = BitSet<config::_64bit>;
@@ -908,4 +974,12 @@ fn index_cursor_test2(){
     
     let mut iter = seq.iter().move_to(c);
     assert_eq!(iter.next().unwrap(), milestone);
+}
+
+#[test]
+fn empty_block_cursor_clone_regression() {
+    let set = HiSparseBitset::new();
+    let c = IndexCursor::end();
+    let i = set.iter().move_to(c);
+    let _ = i.clone();
 }